backend/requirements.txt
Django==3.2.3
flake8==6.0.0
flake8-isort==6.0.0
Pillow==9.5.0
djangorestframework==3.14.0
djoser==2.2.0
